Organize folder structure a bit more
