+++ src/core/Elsa.Core/Elsa.Core.csproj
@@ -48,4 +48,8 @@
         <ProjectReference Include="..\Elsa.Abstractions\Elsa.Abstractions.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Providers" />
+    </ItemGroup>
+
 </Project>
