+++ src/core/Elsa.Core/ElsaOptions.cs
@@ -7,6 +7,8 @@ using Elsa.Persistence;
 using Elsa.Persistence.InMemory;
 using Elsa.Serialization;
 using Elsa.Services;
+using Elsa.Services.Dispatch;
+using Elsa.Services.Messaging;
 using Microsoft.Extensions.DependencyInjection;
 using Newtonsoft.Json;
 using NodaTime;
