+++ src/core/Elsa.Core/Extensions/ElsaServiceCollectionExtensions.cs
@@ -5,7 +5,6 @@ using Elsa.Activities.ControlFlow;
 using Elsa.Activities.Signaling;
 using Elsa.Activities.Signaling.Services;
 using Elsa.Activities.Workflows;
-using Elsa.ActivityTypeProviders;
 using Elsa.Bookmarks;
 using Elsa.Builders;
 using Elsa.Consumers;
@@ -13,7 +12,6 @@ using Elsa.Converters;
 using Elsa.Decorators;
 using Elsa.Design;
 using Elsa.Dispatch;
-using Elsa.Dispatch.Consumers;
 using Elsa.Events;
 using Elsa.Expressions;
 using Elsa.Handlers;
@@ -22,14 +20,21 @@ using Elsa.Mapping;
 using Elsa.Metadata;
 using Elsa.Persistence;
 using Elsa.Persistence.Decorators;
+using Elsa.Providers.ActivityTypes;
+using Elsa.Providers.Workflows;
 using Elsa.Runtime;
 using Elsa.Serialization;
 using Elsa.Serialization.Converters;
 using Elsa.Services;
+using Elsa.Services.Bookmarks;
+using Elsa.Services.Dispatch.Consumers;
+using Elsa.Services.Locking;
+using Elsa.Services.Messaging;
+using Elsa.Services.Triggers;
+using Elsa.Services.WorkflowContexts;
 using Elsa.Services.Workflows;
 using Elsa.StartupTasks;
 using Elsa.Triggers;
-using Elsa.WorkflowProviders;
 using MediatR;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Newtonsoft.Json;
