+++ src/core/Elsa.Core/Services/Triggers/TriggerFinder.cs
@@ -3,8 +3,9 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Elsa.Bookmarks;
+using Elsa.Triggers;
 
-namespace Elsa.Triggers
+namespace Elsa.Services.Triggers
 {
     public class TriggerFinder : ITriggerFinder
     {
