+++ src/core/Elsa.Core/Services/Triggers/TriggerStore.cs
@@ -2,9 +2,10 @@
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Elsa.Triggers;
 using Microsoft.Extensions.Caching.Memory;
 
-namespace Elsa.Triggers
+namespace Elsa.Services.Triggers
 {
     public class TriggerStore : ITriggerStore
     {
