+++ src/server/Elsa.Server.Hangfire/Extensions/ServiceCollectionExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using Elsa.Dispatch;
 using Elsa.Server.Hangfire.Dispatch;
+using Elsa.Services.Dispatch;
 
 namespace Elsa.Server.Hangfire.Extensions
 {
