+++ src/server/Elsa.Server.Orleans/Extensions/ServiceCollectionExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using Elsa.Dispatch;
 using Elsa.Server.Orleans.Dispatch;
+using Elsa.Services.Dispatch;
 
 namespace Elsa.Server.Orleans.Extensions
 {
