+++ src/servicebus/Elsa.Rebus.AzureServiceBus/Extensions/ElsaOptionsExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Elsa.Services;
+using Elsa.Services.Messaging;
 using Microsoft.Azure.ServiceBus.Primitives;
 using Rebus.Config;
