+++ src/servicebus/Elsa.Rebus.RabbitMq/Extensions/ElsaOptionsBuilderExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using Elsa.Services;
+using Elsa.Services.Messaging;
 using Rebus.Config;
 
 // ReSharper disable once CheckNamespace
