+++ test/unit/Elsa.UnitTests/Triggers/TriggersForBlueprintsProviderTests.cs
@@ -6,6 +6,7 @@ using AutoFixture.Xunit2;
 using Elsa.Models;
 using Elsa.Services;
 using Elsa.Services.Models;
+using Elsa.Services.Triggers;
 using Elsa.Testing.Shared;
 using Elsa.Testing.Shared.AutoFixture.Attributes;
 using Moq;
